+++ mud/things/dungeon.js
@@ -1,4 +1,4 @@
-const RedisGraphShim = require('../redis-graph-shim')
+const RedisGraphShim = require('../data/redis-graph-shim')
 
 const Room = require('./room')
